+++ scripts/build_antithesis_images.sh
@@ -17,13 +17,13 @@ if [[ "${TEST_SETUP}" != "node" && "${TEST_SETUP}" != "xsvm" ]]; then
 fi
 
 # Directory above this script
-AVALANCHE_PATH=$( cd "$( dirname "${BASH_SOURCE[0]}" )"; cd .. && pwd )
+LUX_PATH=$( cd "$( dirname "${BASH_SOURCE[0]}" )"; cd .. && pwd )
 
-source "${AVALANCHE_PATH}"/scripts/constants.sh
-source "${AVALANCHE_PATH}"/scripts/git_commit.sh
+source "${LUX_PATH}"/scripts/constants.sh
+source "${LUX_PATH}"/scripts/git_commit.sh
 
 # Import common functions used to build images for antithesis test setups
-source "${AVALANCHE_PATH}"/scripts/lib_build_antithesis_images.sh
+source "${LUX_PATH}"/scripts/lib_build_antithesis_images.sh
 
 # Specifying an image prefix will ensure the image is pushed after build
 IMAGE_PREFIX="${IMAGE_PREFIX:-}"
@@ -41,7 +41,7 @@ GO_VERSION="$(go list -m -f '{{.GoVersion}}')"
 # Helper to simplify calling build_builder_image for test setups in this repo
 function build_builder_image_for_node {
   echo "Building builder image"
-  build_antithesis_builder_image "${GO_VERSION}" "antithesis-node-builder:${IMAGE_TAG}" "${AVALANCHE_PATH}" "${AVALANCHE_PATH}"
+  build_antithesis_builder_image "${GO_VERSION}" "antithesis-node-builder:${IMAGE_TAG}" "${LUX_PATH}" "${LUX_PATH}"
 }
 
 # Helper to simplify calling build_antithesis_images for test setups in this repo
@@ -57,35 +57,35 @@ function build_antithesis_images_for_node {
     echo "Building images for ${test_setup}"
   fi
   build_antithesis_images "${GO_VERSION}" "${image_prefix}" "antithesis-${test_setup}" "${IMAGE_TAG}" "${IMAGE_TAG}" \
-                          "${AVALANCHE_PATH}/tests/antithesis/${test_setup}/Dockerfile" "${uninstrumented_node_dockerfile}" \
-                          "${AVALANCHE_PATH}" "${node_only}" "${git_commit}"
+                          "${LUX_PATH}/tests/antithesis/${test_setup}/Dockerfile" "${uninstrumented_node_dockerfile}" \
+                          "${LUX_PATH}" "${node_only}" "${git_commit}"
 }
 
 if [[ "${TEST_SETUP}" == "node" ]]; then
   build_builder_image_for_node
 
   echo "Generating compose configuration for ${TEST_SETUP}"
-  gen_antithesis_compose_config "${IMAGE_TAG}" "${AVALANCHE_PATH}/tests/antithesis/node/gencomposeconfig" \
-                                "${AVALANCHE_PATH}/build/antithesis/node"
+  gen_antithesis_compose_config "${IMAGE_TAG}" "${LUX_PATH}/tests/antithesis/node/gencomposeconfig" \
+                                "${LUX_PATH}/build/antithesis/node"
 
-  build_antithesis_images_for_node "${TEST_SETUP}" "${IMAGE_PREFIX}" "${AVALANCHE_PATH}/Dockerfile" "${NODE_ONLY:-}"
+  build_antithesis_images_for_node "${TEST_SETUP}" "${IMAGE_PREFIX}" "${LUX_PATH}/Dockerfile" "${NODE_ONLY:-}"
 else
   build_builder_image_for_node
 
   # Only build the node node image to use as the base for the xsvm image. Provide an empty
   # image prefix (the 1st argument) to prevent the image from being pushed
   NODE_ONLY=1
-  build_antithesis_images_for_node node "" "${AVALANCHE_PATH}/Dockerfile" "${NODE_ONLY}"
+  build_antithesis_images_for_node node "" "${LUX_PATH}/Dockerfile" "${NODE_ONLY}"
 
   # Ensure node and xsvm binaries are available to create an initial db state that includes subnets.
   echo "Building binaries required for configuring the ${TEST_SETUP} test setup"
-  "${AVALANCHE_PATH}"/scripts/build.sh
-  "${AVALANCHE_PATH}"/scripts/build_xsvm.sh
+  "${LUX_PATH}"/scripts/build.sh
+  "${LUX_PATH}"/scripts/build_xsvm.sh
 
   echo "Generating compose configuration for ${TEST_SETUP}"
-  gen_antithesis_compose_config "${IMAGE_TAG}" "${AVALANCHE_PATH}/tests/antithesis/xsvm/gencomposeconfig" \
-                                "${AVALANCHE_PATH}/build/antithesis/xsvm" \
-                                "LUXD_PATH=${AVALANCHE_PATH}/build/node AVAGO_PLUGIN_DIR=${AVALANCHE_PATH}/build/plugins"
+  gen_antithesis_compose_config "${IMAGE_TAG}" "${LUX_PATH}/tests/antithesis/xsvm/gencomposeconfig" \
+                                "${LUX_PATH}/build/antithesis/xsvm" \
+                                "LUXD_PATH=${LUX_PATH}/build/node LUXD_PLUGIN_DIR=${LUX_PATH}/build/plugins"
 
-  build_antithesis_images_for_node "${TEST_SETUP}" "${IMAGE_PREFIX}" "${AVALANCHE_PATH}/vms/example/xsvm/Dockerfile"
+  build_antithesis_images_for_node "${TEST_SETUP}" "${IMAGE_PREFIX}" "${LUX_PATH}/vms/example/xsvm/Dockerfile"
 fi